refactor(harvest): Extract AccountFields from AccountForm 🛠
